Fix random bugs.

+++ Classes/Weapon/Weapon.cpp
@@ -1,6 +1,6 @@
 #include "Weapon.h"
 
-void Weapon::weaponInit(float speed, int attack, int state, int type) {
+void Weapon::weaponInit(float speed, float attack, int state, int type) {
 	setWeaponSpeed(speed);
 	setWeaponAttack(attack);
 	setWeaponState(state);
